// src/app/models/starship.ts
export class Starship {
    name: string;
    model: string;
    MGLT: string;
    cargo_capacity: string;
    cost_in_credits: string;
    crew: string;
    starship_class: string;
    hyperdrive_rating: string;
    manufacturer: string;
    passengers: string;
    max_atmosphering_speed: string;
    length: string;
}